Factored code to it's own struct
