+++ edgelet/edgelet-kube/src/runtime.rs
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-#![allow(unused_variables, dead_code)] // todo remove after
+#![allow(unused_variables, unused_imports, dead_code)] // todo remove after
 
 use std::cell::RefCell;
 use std::sync::{Arc, Mutex};
@@ -10,6 +10,7 @@ use failure::Fail;
 use futures::future::Either;
 use futures::prelude::*;
 use futures::{future, stream, Async, Future, Stream};
+use hyper::header::HeaderValue;
 use hyper::service::Service;
 use hyper::{header, Body, Chunk as HyperChunk, Request};
 use log::Level;
@@ -28,8 +29,7 @@ use crate::convert::{
     auth_to_image_pull_secret, pod_to_module, spec_to_deployment, spec_to_service_account,
 };
 use crate::error::{Error, ErrorKind, Result};
-use crate::module::KubeModule;
-use hyper::header::HeaderValue;
+use crate::module::{CreateModule, KubeModule};
 
 #[derive(Clone)]
 pub struct KubeModuleRuntime<T, S> {
@@ -101,156 +101,6 @@ impl<T, S> KubeRuntimeData for KubeModuleRuntime<T, S> {
     fn management_uri(&self) -> &Url {
         &self.management_uri
     }
-}
-
-fn create_service_account<T, S>(
-    runtime: &KubeModuleRuntime<T, S>,
-    module: ModuleSpec<DockerConfig>,
-) -> Box<dyn Future<Item = (), Error = Error> + Send>
-where
-    T: TokenSource + Clone + Send + 'static,
-    S: Send + Service + 'static,
-    S::ReqBody: From<Vec<u8>>,
-    S::ResBody: Stream,
-    Body: From<S::ResBody>,
-    <S::ResBody as Stream>::Item: AsRef<[u8]>,
-    <S::ResBody as Stream>::Error: Into<Error>,
-    <S::ResBody as Stream>::Error: Into<KubeClientError>,
-    S::Error: Into<KubeClientError>,
-    S::Future: Send,
-{
-    let f = spec_to_service_account(runtime, &module)
-        .map_err(Error::from)
-        .map(|new_service_account| {
-            let client_copy = runtime.client.clone();
-            let namespace_copy = runtime.namespace().to_owned();
-            runtime
-                .client
-                .lock()
-                .expect("Expected lock error")
-                .borrow_mut()
-                .list_service_accounts(
-                    runtime.namespace(),
-                    Some(module.name()),
-                    Some(&runtime.device_hub_selector),
-                )
-                .map_err(Error::from)
-                .and_then(move |service_accounts| {
-                    if let Some(current_service_account) =
-                        service_accounts.items.into_iter().find(|service_account| {
-                            service_account.metadata.as_ref().map_or(false, |meta| {
-                                meta.name.as_ref().map_or(false, |n| *n == module.name())
-                            })
-                        })
-                    {
-                        // found service account, if the service account doesn't match, replace it
-                        if current_service_account == new_service_account {
-                            Either::A(Either::A(future::ok(())))
-                        } else {
-                            let fut = client_copy
-                                .lock()
-                                .expect("Unexpected lock error")
-                                .borrow_mut()
-                                .replace_service_account(
-                                    namespace_copy.as_str(),
-                                    module.name(),
-                                    &new_service_account,
-                                )
-                                .map_err(Error::from)
-                                .map(|_| ());
-                            Either::A(Either::B(fut))
-                        }
-                    } else {
-                        // not found - create it
-                        let fut = client_copy
-                            .lock()
-                            .expect("Unexpected lock error")
-                            .borrow_mut()
-                            .create_service_account(namespace_copy.as_str(), &new_service_account)
-                            .map_err(Error::from)
-                            .map(|_| ());
-                        Either::B(fut)
-                    }
-                })
-        })
-        .flatten();
-
-    Box::new(f)
-}
-
-fn create_deployment<T, S>(
-    runtime: &KubeModuleRuntime<T, S>,
-    module: ModuleSpec<DockerConfig>,
-) -> Box<dyn Future<Item = (), Error = Error> + Send>
-where
-    T: TokenSource + Clone + Send + 'static,
-    S: Send + Service + 'static,
-    S::ReqBody: From<Vec<u8>>,
-    S::ResBody: Stream,
-    Body: From<S::ResBody>,
-    <S::ResBody as Stream>::Item: AsRef<[u8]>,
-    <S::ResBody as Stream>::Error: Into<Error>,
-    <S::ResBody as Stream>::Error: Into<KubeClientError>,
-    S::Error: Into<KubeClientError>,
-    S::Future: Send,
-{
-    let f = spec_to_deployment(runtime, &module)
-        .map_err(Error::from)
-        .map(|(name, new_deployment)| {
-            let client_copy = runtime.client.clone();
-            let namespace_copy = runtime.namespace().to_owned();
-            runtime
-                .client
-                .lock()
-                .expect("Unexpected lock error")
-                .borrow_mut()
-                .list_deployments(
-                    runtime.namespace(),
-                    Some(&name),
-                    Some(&runtime.device_hub_selector),
-                )
-                .map_err(Error::from)
-                .and_then(move |deployments| {
-                    if let Some(current_deployment) =
-                        deployments.items.into_iter().find(|deployment| {
-                            deployment.metadata.as_ref().map_or(false, |meta| {
-                                meta.name.as_ref().map_or(false, |n| *n == name)
-                            })
-                        })
-                    {
-                        // found deployment, if the deployment found doesn't match, replace it.
-                        if current_deployment == new_deployment {
-                            Either::A(Either::A(future::ok(())))
-                        } else {
-                            let fut = client_copy
-                                .lock()
-                                .expect("Unexpected lock error")
-                                .borrow_mut()
-                                .replace_deployment(
-                                    namespace_copy.as_str(),
-                                    name.as_str(),
-                                    &new_deployment,
-                                )
-                                .map_err(Error::from)
-                                .map(|_| ());
-                            Either::A(Either::B(fut))
-                        }
-                    } else {
-                        // Not found - create it.
-                        let fut = client_copy
-                            .lock()
-                            .expect("Unexpected lock error")
-                            .borrow_mut()
-                            .create_deployment(namespace_copy.as_str(), &new_deployment)
-                            .map_err(Error::from)
-                            .map(|_| ());
-                        Either::B(fut)
-                    }
-                })
-        })
-        .into_future()
-        .flatten();
-    Box::new(f)
 }
 
 impl<T, S> KubeModuleRuntime<T, S> {
@@ -335,6 +185,14 @@ impl<T, S> KubeModuleRuntime<T, S> {
             management_uri,
             device_hub_selector,
         })
+    }
+
+    pub fn client(&self) -> Arc<Mutex<RefCell<KubeClient<T, S>>>> {
+        self.client.clone()
+    }
+
+    pub fn device_hub_selector(&self) -> &str {
+        &self.device_hub_selector
     }
 }
 
@@ -424,7 +282,7 @@ where
 impl<T, S> ModuleRuntime for KubeModuleRuntime<T, S>
 where
     T: TokenSource + Clone + Send + 'static,
-    S: Send + Service + 'static,
+    S: Service + Clone + Send + 'static,
     S::ReqBody: From<Vec<u8>>,
     S::ResBody: Stream,
     Body: From<S::ResBody>,
@@ -441,7 +299,7 @@ where
     type Chunk = Chunk;
     type Logs = Logs;
 
-    type CreateFuture = Box<dyn Future<Item = (), Error = Self::Error> + Send>;
+    type CreateFuture = Box<dyn Future<Item = (), Error = Error> + Send>;
     type GetFuture =
         Box<dyn Future<Item = (Self::Module, ModuleRuntimeState), Error = Self::Error> + Send>;
     type InitFuture = Box<dyn Future<Item = (), Error = Self::Error> + Send>;
@@ -461,8 +319,7 @@ where
     }
 
     fn create(&self, module: ModuleSpec<Self::Config>) -> Self::CreateFuture {
-        let f = create_service_account(self, module);
-        Box::new(f)
+        CreateModule::new(self.clone(), module)
     }
 
     fn get(&self, _id: &str) -> Self::GetFuture {
